pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bigbookapi",
    "flask>=3.1.2",
    "flask-restful>=0.3.10",
    "omdb>=0.10.1",
    "pyodbc>=5.3.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "tmdbv3api>=1.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[tool.uv.sources]
bigbookapi = { git = "https://github.com/ddsky/big-book-api-clients.git", subdirectory = "python" }
